# docs/swagger.yaml
definitions:
  controllers.InputUpdatePassword:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  controllers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterInputSekolah:
    properties:
      nama:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - nama
    - password
    - username
    type: object
  controllers.RegisterInputSiswa:
    properties:
      kode_verifikasi_pendaftaran:
        type: string
      nama:
        type: string
      nisn:
        type: string
      password:
        type: string
      sekolah_id:
        type: integer
      username:
        type: string
    required:
    - kode_verifikasi_pendaftaran
    - nama
    - nisn
    - password
    - sekolah_id
    - username
    type: object
  models.Angkatan:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nama:
        type: string
      sekolah_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Users:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nama:
        type: string
      password:
        type: string
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  sekolahController.InputAngakatan:
    properties:
      nama:
        type: string
      sekolah_id:
        type: integer
    required:
    - nama
    - sekolah_id
    type: object
...
  sekolahController.UpdateInputAngakatan:
    properties:
      nama:
        type: string
    required:
    - nama
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /daftar-sekolah:
    post:
      description: Daftar Akun Sekolah hanya oleh super admin.
      parameters:
      - description: body unutk daftar akun Sekolah
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInputSekolah'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Daftar Akun Sekolah (akun, kontrol sekolah -> menu-menu, pendaftaran
        pendidik/siswa).
      tags:
      - Auth
  /daftar-siswa:
    post:
      description: Daftar Akun Siswa table yang terdaftar ada 4 yaitu table users
        sebagai akun, table siswa sebagai biodata, table prestasi sebagai manajemen
        prestasi siswa, table asal sekolah sebagai manajemen asal sekolah siswa.
      parameters:
      - description: body untuk daftar akun siswa
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInputSiswa'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Daftar Akun Siswa (akun, biodata, asal sekolah, prestasi, beasiswa).
      tags:
      - Auth
  /ganti-password/{username}:
    patch:
      description: Menggati Password akun berdasarkan username.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: body untuk ganti password
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.InputUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Users'
      summary: Menggati Password akun.
      tags:
      - Auth
  /login:
    post:
      description: Logging in to get jwt token to access admin or user api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as as user.
      tags:
      - Auth
...
  /sekolah/{id}/angkatan:
    get:
      description: Mendapatkan semua data angkatan Berdasarkan ID Sekolah.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Angkatan'
            type: array
      summary: mendapatkan data angkatan berdasarkan ID Sekolah.
      tags:
      - Sekolah -> Angkatan
...
  /sekolah/angkatan:
    post:
      description: Membuat data Angkatan baru.
      parameters:
      - description: the body to create a new Angkatan
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/sekolahController.InputAngakatan'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Angkatan'
      security:
      - BearerToken: []
      summary: Membuat data Angkatan barun.
      tags:
      - Sekolah -> Angkatan
  /sekolah/angkatan/{id}:
    delete:
      description: Menghapus Sebuah data angkatan  berdasarkan id.
      parameters:
      - description: Angkatan id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Menghapus data angkatan.
      tags:
      - Sekolah -> Angkatan
    get:
      description: Mendapatkan data angkatan berdasarkan id.
      parameters:
      - description: Angkatan id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Angkatan'
      summary: Mendapatkan data angkatan.
      tags:
      - Sekolah -> Angkatan
    patch:
      description: memperbaharui data angkatan berdasarkan id.
      parameters:
      - description: Angkatan id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update Angkatan
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/sekolahController.UpdateInputAngakatan'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Angkatan'
      security:
      - BearerToken: []
      summary: Memperbarui data Angkatan.
      tags:
      - Sekolah -> Angkatan
...
swagger: "2.0"